Added update and POST methods
The POST method should update (not append) data to the store.
Fixed a couple of poorly written test cases that didn't recognize the
state of changing data. I was also calling new on the service. I
moved it to a singleton to get a single instance of the shared data
state.