Corrected Usage of PUT/POST
Both Put and Post are completely valid to create new data. In fact
it's a falacy to say, that PUT/POST responds to the CREATE and UPDATE
in CRUD. PUT is used when you are defining where and what is going
at a location. So inserting B123 with the data is a PUT. It is also
idempotent (it will not insert two B123). Post is used when you are
allowing the server to decide where the data you are inserting is
going. It is NOT idempotent. Calling post twice will insert two
records at two different locations. In the bookstore example, POST
should not be used. The server never creates its own location.